+++ src/main/java/com/clothes/services/PriceService.java
@@ -4,6 +4,8 @@ import com.clothes.repositories.PriceRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+
 
 @Service
 public class PriceService {
@@ -11,11 +13,11 @@ public class PriceService {
     @Autowired
     private PriceRepository priceRepository;
 
-    public double findMaxPrice() {
+    public BigDecimal findMaxPrice() {
         return  priceRepository.findByMaxPrice();
     }
 
-    public double findMimPrice() {
+    public BigDecimal findMimPrice() {
         return  priceRepository.findByMinPrice();
     }
